Split off synthesis to its own module
